- Remove server dependency on KDELibs, switching to our own functionality
Till: I initially thought about doing it as a singleton because, as you wrote,
it is more convenient. However, since the class is shared between server and
libakonadi, we can not just use KDE's static deleter, so we will need
explicit creation/deletion points. Will be easy to do in the server processes
since they have main(), but I'm not sure where to do this for usage in libakonadi.
Session ctor and dtor maybe?

CCMAIL: adam@kde.org

svn path=/trunk/KDE/kdepim/akonadi/server/; revision=704874