Backport SVN commit 1003796 by krake from trunk:
Fixing check for presence of a list.
If stripLeadingSpaces() puts the position at the end of the input data, we need to read more data.

Actually this should probably be done in a loop if there is more than one leading space, not sure if the protocol allows that though.

svn path=/branches/akonadi/1.2/; revision=1012883