Change how missing ancestors are inserted into the model.
If an ancestor is filtered out, and we wish to show one of its descendants in the model,
the ancestor is not in the model, so it must be fetched.

Now a minimal collection is inserted into the Collection tree and a job for the actual ancestor collection is started.

svn path=/trunk/KDE/kdepimlibs/; revision=1015117