Make the EntityTreeModel take a ChangeRecorder instead of a Monitor.
This is necessary to resolve a race condition as described in the
mailing list a few weeks ago.

This is backwards incompatible, but it's @since 4.4, so that's ok.

I have s/Monitor/ChangeRecorder/ patches for kdepim, playground/pim
and Mailody. If anything else is broken by this, just make the same change.

CCMAIL: kde-pim@kde.org

svn path=/trunk/KDE/kdepimlibs/; revision=1029709