Backport from trunk.
r1077216 | tmcguire | 2010-01-19 18:57:13 +0100 (Tue, 19 Jan 2010) | 12 lines

Work around a DBus bug that using a connection from within multiple threads can cause freezes.
In this case, that connection was the session bus, on which we called isServiceRegistered.

The workaround: Just don't check if the search service is available, just call the methods. They'll
fail properly if the service is not there.

Volker, can you backport this fix to older Akonadi releases and prepare tarballs, so that we can
notify the release team about this?

Should fix
CCBUG: 219687

svn path=/branches/akonadi/1.3/; revision=1077436