Start of the remote debugging server.
It forwards the local Akonadi socket to a TCP socket, to allow
connecting akonadiconsole to an Akonadi server running on e.g. a mobile
device. To make this work we not only need to forward the Akonadi socket
though, but also D-Bus and Nepomuk. Therefore this is not build into the
Akonadi server itself (which would have been a lot easier), but a
separate tools that will handle all those sockets eventually.

svn path=/trunk/kdesupport/akonadi/; revision=1096771