use single query in ItemRetriever
This patch makes the ItemRetriever use a single SQL query with LeftJoins to achieve the same as before.

This won't reduce the overall query time, since the same amount of items will be retrieved.
But imo it makes the code much cleaner and also enables us to do some optimizations in the future:

- data.isNull directly in the query
- item retrieval from the resource for lists of items, not only single items

Review: http://reviewboard.kde.org/r/3876/

svn path=/trunk/kdesupport/akonadi/; revision=1122335