fix collation and charsets in mysql database
Akonadi works with UTF internally but the MySQL database uses latin1 for collation and default charsets, which this patch should fix.

I have no idea how to do that for SQLite or PSQL, nor whether it's actually a problem there. But for MySQL this also
fixes the test cases that emit 'illegal mix of collations' error messages.

review: http://reviewboard.kde.org/r/3877/

svn path=/trunk/kdesupport/akonadi/; revision=1122604