refactor JOIN in QueryBuilder, esp. in regard to INNER JOIN
Now the QueryBuilder takes it's main table in the Ctor as it's required for all queries anyways.
All joins, even INNER JOINs are handled via the new addJoin() method that takes a JoinType argument.

- LEFT JOIN is only supported for SELECT queries
- INNER JOIN is only supported for SELECT and UPDATE queries

This change to the handling of INNER JOINs is required, since the ',' as shortcut for an INNER JOIN
cannot be mixed with LEFT JOINs, while the explicit version can be. See also:

http://dev.mysql.com/doc/refman/5.0/en/join.html

The part about 'unknown column' when using ',' for INNER JOINs.

Note: Since INNER JOINs are more restrictive than LEFT JOINs, the former
will take precedence over the latter. Joining one and the same table twice
is not supported, instead the JoinConditions will be merged.

based on: http://reviewboard.kde.org/r/3907/

svn path=/trunk/kdesupport/akonadi/; revision=1125001