optimize: only query for the empty parts in the parttable and request an update for them
before we went through all parts of a collection and loaded the data (maybe from disk)
and looked whether the data was null or non existing.

now we use the power of the left join to find items without any fitting part
(which triggers update of all requested parts). alternatively we cope with
invalidated cache entries (which have datasize == 0).

svn path=/trunk/kdesupport/akonadi/; revision=1126817