Reselect resource after session reconnect.
Every resource executes a ResourceSelectJob on startup to
tell the server that this session belongs to an resource.
If the server was restarted (e.g. after a server crash), the session
had been reconnected, but the ResourceSelectJob not been re-executed.
This lead to failing of resource-specific jobs in the new session.

svn path=/trunk/KDE/kdepimlibs/; revision=1207333