Recover from loss of resource information in the database.
This recreates missing entries in the resource table during startup, which
means that we can now automatically recover from a total loss of the
database if we still have a valid configuration. Of course, not yet synced
changes are lost when deleting the database, so don't actually try that
unless you know what you are doing.