Reverse the order of FETCH responses from ascending to descending.
While FETCH does not guarantee any particular order, we need sorted results
internally in the server, but the actual order doesn't really matter there.
However, this considerably increases the perceived responsiveness of any
model that shows the most recent entries first (such as the KMail message
list). Of course the full loading takes just as long as before and you can
easily construct pathological cases where it has the opposite effect, but
on average it seems to improve things.