Improve updates in mobile UI mode
In mobile UI mode we do not insert new items as soon
as ItemFetchJob reports them via the itemsRetrieved() signal,
but wait until the ItemFetchJob has finished and then we'll
insert all items in one go. This prevents rowsInserted() signals
of the model, which will cause unnecessary (and expensive) updates
in the proxy stack above.

svn path=/trunk/KDE/kdepimlibs/; revision=1211207