Make ETM fetch empty collections at least once
When collections are empty because they have just been created
(synchronizeCollectionTree or equivalent operation), an application
using EntityTreeModel will not fetch their contents, assuming they
are truly empty (contain no items). This is done for optimization
purposes.

This patch make ETM only skip collections which are known to be
empty. We can't just check the statistics for the optimization
because the statistics will report zero until something is requested
if the resource simply has not been sync'd yet.

BUG: 271913
REVIEW: 101844