Make sure we always use the part id for creating part file names.
This was found by the new overlapping part checker and seemed to result
in a <0.07% chance of having a part overwrite the file of another one.
Since this will not be written back to the resource normally, the actual
risk of data damage is probably even smaller.
(cherry picked from commit add06052104974356404a0eccf09dc1c21325b0c)