Only mark items as dirty if their payload changed.
Right now we have the problem that items stay dirty if the last change
affected a part that the resource does not store (e.g. attributes). As
a side-effect, they are ignored by the cache cleaner, in order to avoid
data loss of unsaved changes. However, the cache cleaner (as the only
user of the dirty flag) only looks at payload parts anyway, so there is
no point in marking an item as dirty if no payload has been changed.