Use the akonadiItemId property to get the item id.
The assumption that the nepomuk uri is the same as the akonadi uri holds no longer true.
The requested property would ideally be added to the existing query, instead we rely on the caller to request the property for now or use a second query.
(cherry picked from commit e67a9d59790740a3ecaa6c81bfaeeb7fec726fd4)