Avoid a nepomuk round trip by requesting the akonadi item id property manually in all cases.
Normally when a property is asked for as part of at sparql query, the
query engine should extract it from the string and supply it as a
property in the result object. This seems to not work at the moment, so
to make sure we actually get the akonadi item id back from searches, and
avoid having to roundtrip again to get them, let's add the property
request manually, like the stored searches code does.