Set the MySQL binlog format explicitly.
The combination of the transaction isolation level we use and the default
binlog format doesn't work. Newer MySQL versions (>= 5.5 apparently)
automatically switch to a supported format, older ones just fail.

We don't actually enable the binlog, nor do we need it, so I don't really
know why it's on for some users in the first place.

BUG: 295700