Revert "Set the MySQL binlog format explicitly."
This reverts commit 8f6eb20aaa854224a6d4070ee2459290ec431576.

It turns out that MySQL <= 5.5 fails with a binlog_format specified,
but no binlog enabled (which we don't want to enabled). Further
research on why this was failing for a few people before showed
broken mysql.conf files that for some reason contained a section
that enabled the binlog explicitly. If that's more than a handfull
of people affected, we can add a filter in Akonadi that removes
those lines when assembling the config file.