Make the query string column wide enough for Sparql queries.
Ideally we would be using BLOB, but that is a bit of a pain to add (would require to alter the whole mapping of Qt-Datatypes, or use some magic value), and 32000 should be big enough for sane queries.
As an extra there is now a safety check for oversized queries, so we don't get truncated queries executed.