Fix threading issue when retrieving missing IDs.
This prevents the creation of new blocking Nepomuk queries while being in event
loop of blocking query by mSearchService (newEntries signal is connected to
NepomukSearch::hitsAdded(...)). The corresponding code to request missing IDs
for results without IDs is moved to a point where the first blocking query is
finished and it is save to create new queries.

REVIEW: 105599