Make referential action explicit.
So far we relied on the various backends adding those, and we had no way
of defining what the actual actions should be (DELETE doesn't make sense
everywhere).

This doesn't yet fix the inconsistencies in the n:m tables nor does it
correct previously missing/wrong referential actions, that's the next
step.