Make the remaining stuff async.
InvitationHandler uses MailScheduler which is async and was
doing error handling ( or not doing any ) when the MailScheduler
call returned, instead of waiting for the underlying thread/job
to finish.

This code base already suffered a lot: kde3->kde4 port, akonadi
port and now forcing it to be async, so it's a bit of a mess,
therefore I won't export this class, will only be used by
IncidenceChanger.