Make killing running jobs work, by reconnecting the corresponding session.
This is rather brutal and can have side-effects on dependent jobs (e.g.
when using transactions), so use with care (best on separated sessions if
you expect a lot of job canceling). Killing queued jobs that have not
started communicating with the server yet will not cause a session
reconnect.

CCMAIL: dfaure@kde.org