Initial port of Akonadi server to new notifications
Internally, Akonadi server now uses NotificationMessageV2 which can,
for specific operations, handle multiple items.

The code needs to be optimized on several places and compression is
not yet implemented, but legacy clients (i.e. clients that don't
support the new format) work perfectly. There is no client that would
support the new format, so that is not yet tested :-)