Allow to disable the cache verification on retrieval.
This was originally added to counter the cache loss in the fs backend which
back then wasn't understood, but which recently got fixed. It comes at a
rather hefty price though, about 30% of a mail folder listing are
attributed to it.

So, switch it off by default for 1.10 now that the underlying issue got
fixed.