Dont' assert if Incidence has empty UID.
It's not a problem with our software, so dont' crash, it's something
for the user to fix.

We could filter this right off the bat in kcalcore/icalformat_p.cpp
but it's useful to have them in memory because i'm implementing
a dialog in korg to automatically fix broken incidences.

BUG: 320329