Don't compress ModifyFlags into Modify operations
In a situation that ModifyFlags was emitted right after
Modify operation (for the same items), the ModifyFlags
operation was merged into the Modify operation, and thus
effectively lost, because addedFlags and removedFlags
fields are ignored in Modify operation.

Fixes a bug that caused emails that were marked as read (usually
during sync) to revert to unread state after next sync.