Save as much as 60MB of heap space.
This patch clears the query cache if it's unused for more than 30 seconds.

Tested with two clients connected (korg and korgac).
Massif pointed out 60MB of heap allocated to the QueryCache,
massive blobs in libmysql.

The QueryCache was created to speed up mass operations,
like deleting all items from a collection, after that,
there's no use in keeping it around.

Even if another mass operation arrives, it's just a few milliseconds
to prepare the first query so it can be used for all following queries
of this new mass-operation.

My akonadiserver is now running at 17MB when using jemalloc.

( with glibc malloc it's a bit more random, due to glibc's
complicated policies to return memory to OS, sometimes it goes
down to 17MB too, other times stays at 80MB ).

REVIEW: 111203