Store filename of external payload without path
Store only filename instead of the whole path in parttable. The path is always the same
so this can save quite a lot of disk space and since server knows the path it can easily
prepend it to the filename before sending it to clients.

When client announces NOPAYLOADPATH capability, the path is not prepended to filename,
which reduces size of replies a bit.

REVIEW: 111541