Support for Gid (globally unique identifier).
This patch adds a new column for the gid of an item to the pimitemtable.
Support for:
* append commands
* store command
* fetch command

We ignore the collection/resource context for now.
Otherwise fetch jobs may silently fail due to a selected collection
(apparently the selection is not reset by a new fetchjob).

Increased the protocol version to 34.