Added support for GID.
This patch adds support for a globally unique identifier for items.
In particular it allows to set the GID from ItemCreate/ModifyJobs,
to refer to items using the GID in fetch/modify jobs, and to retrieve
the GID in fetch jobs using the itemfetch scope.

The GidExtractorInterface can be implemented by serializer plugins to
extract the GID from the payload of an item.
A typical example is the uid of an ical object.

By using this facility akonadi supports efficient lookups by gid.

This feature requires the server protocol version 34.

REVIEW: 110648