Fix creating items with a single external payload.
This was removing the just inserted part from the db again, but keeping the
external payload file around. This cannot possibly have ever worked...

We probably didn't notice this so far since emails are multi-part and don't
hit this path at all, and everything else is usually too small for the
external storage threshold. This was found by the unit tests for
ItemCreateJob in kdepimlibs.

REVIEW: 111819