Change the notifications filter to match the one from Akonadi::Monitor
The reverse filtering (i.e. getting list of sources that accept
given notification) cannot be mapped 1:1 to the original behavior
of Akonadi::Monitor easily while keeping the code readable and
causes various subtle bugs in filtering out (or not filtering)
certain notifications. Therefor we revert to the original approach
of asking each source whether it accepts given notification.

In the end it is probably faster than the previous approach
because it does only simple comparisions instead of uniting sets.

And except for other things it fixes problem with endless item
fetch loop when resource cannot retrieve all requested payload parts
(emails without a body for instance), an example of the subtle bug
mentioned above.