First take on PartTable migration
This patch is able to migrate PartTable on MySQL backend, but because
ALTER TABLE query on MySQL copies the table, it takes ages to complete.

One of the changes done in this code is delayed index and FK constraints
creation. That means that after DbInitializer checks whether all tables
and columns are present it will not add the indexes but instead DbUpdater
is called. It will perform the update, populate new tables and columns,
etc. and after that the DbInitializer is called again to check and update
the indexes. This prevents SQL errors when creating UNIQUE INDEX on a table
with a new column, where all values are set to default value, thus not
unique. Same applies for foreign keys.

PostgreSQL fails so far, because it can't add a new column to a non-empty
table without specifying default value. Specifying default value will not
work either because the foreign key constraint pointing to (so far empty)
PartTypeTable will fail (PartTypeTable has no record with id 0).

SQLite has not been tested.