Don't say we can fetchMore if we really can't.
Since we purge all items, the usual checks apply. This extra check was meant to get items updated
that were still cached, but potentially outdated since we didn't receive any updates due to the collection not being monitored.

Due to the "always true if !monitored", it would be possible to endlessly refetch items of a non monitored collection,
which is not what we want.