Change path of org.freedesktop.Akonadi.Preprocessor interface
This is a change due to way how client-side libraries are implemented. They
cannot register the Preprocessor object into '/' path, so we use
'/Preprocessor' path instead. We can't fix this easily on client side
without breaking public API too much.

It is also more consistent with other extensions, like Search, which is
in '/Search' path.