SQLite: Handle joins in UPDATE queries
Since SQLite does not directly support JOIN expression in UPDATE or UPDATE on
multiple tables at once, we have to convert the JOIN into a subquery in WHERE
condition.

The conversion is far from perfect, but it passes the unit-tests based on the
query that does this (I think there's only one such query in Akonadi that does
this, see FetchHelper::updateItemAccessTime())

This fixes updating atime on items from virtual collections when using SQLite
backend.

BUG: 305543
FIXED-IN: 1.11