Only emit the dataChanged signal for fetched collections.
For item fetches the collection hasn't changed and row inserted notifications are
used to signal new rows. This fixes and endless fetch loop caused by the
proxy models on top accidentally purging the fetched items after getting
a dataChanged signal, resulting in the items being fetched again, resulting in
a data changed signal, etc.
While this doesn't fix the incorrect behaviour of the proxy models, it still
fixes the incorrect behaviour of the ETM, and resolves the endless loop at least.