Extract FetchScope from FetchHelper so that we can use it independently
We parse the fetch scope and store it in FetchScope, which then can
be cached and can be passed to FetchHelper multiple times, which was
otherwise not possible, as FetchHelper would always try to parse it
from stream.

Thanks to it we can now send results of SEARCH command as they arrive,
instead of waiting for all sources to finish.