Emit escaped identifier in NotificationManager::{un,}subscribed() signals
NotificationManager escapes the identifier by replacing all non-alphanumeric
characters by underscore, so we should emit the escaped identifier instead of
the original one in the (un)subscribed() signals, as clients might be using
the identifier to subscribe to monitor's changes on DBus (Akonadi Console does
for example).

In KF5 I'd like to change the (un)subscribed() singal to have QDBusPath of the
new subscriber instead of just a string argument.