Fix parsing of .desktop files entries with comma in value
We are using QSettings to parse .desktop files. QSettings however treats
a value with comma as a QStringList which is not compatible with KConfig's
behavior, because KConfig treats everything as a string by default.

This patch adds a small workaround for this QSettings behaviour and joins
the list back to a single string when a string value is expected.

For translators this means that Comment entries in Akonadi agents' .desktop
files don't have to be enclosed in double quotes anymore.

BUG: 330010
FIXED-IN: 1.11.1