Prevent main SEARCH parser from parsing fetch scope
Fetch scope is parsed separately, so once we reach "QUERY" parameter in the
main parsing loop, we break, otherwise next iteration hits the fetch scope and
throws "Invalid parameter".

This is sort of a hack since it assumes client will always send QUERY as the
last parameter before fetch scope. We need a better detection...