Fix SEARCH_RESULT command leaving stream parser dirty
When ImapStreamParser::readSequenceSet() hits ')', it returns, leaving
stream parser positioned on the ')' character, so next command on the
session will be ")COMMAND". We fix this by reading the one single
character from stream parser after readSequenceSet() returns, so that
the parser is correctly positioned beyond the end of the command.