Use scheduling in CacheCleaner and IntervalCheck
Instead of repeatadly querying all collections and checking, whether
collection should be synchronized or cleared, we query all collections
only on start and then we just set a QTimer to timeout when a nearest
collection should expire.

This should decrease disk wakeups, especially in IntervalCheck, because
is does only little SQL now, while before it was checking all collections
every minute or so.

New collections or changes in collection cache policies are delivered to
CacheCleaner and IntervalCheck from NotificationCollector.