Handle database transaction deadlocks
Two concurrent database transactions can cause a database deadlock under
certain conditions. Both MySQL and PostgreSQL can detect the deadlock and
will automatically rollback one of the offending transactions. Very often
we've seen atime update in ItemRetriever to cause a deadlock with a
large transaction updating or inserting items, especially from maildir
resource.

According to documentation of both databases, transaction deadlocks are
nothing dangerous and applications should be able to handle them by trying
to replay the rolled-back transaction again.

For that reason we now store all QSqlQueries executed within a single
transaction into DataStore, so that in case of error we can ask DataStore
to execute them all again within a new transaction. After each transaction
is committed or rolled back, DataStore will clear it's query cache.

Queries executed outside transactions are not recorded. Also when using
SQLite, the queries are not recorded either, because SQLite does not support
concurrent transaction and we serialize them through a global lock in
DataStore, so no deadlock can occur.

This commit does not solve the actual problem that we somewhere generate a
huge and slow transaction that causes the deadlock but makes Akonadi able
to handle it without errors. So far I was not able to find where the big
transaction comes from.