Allow FETCH of items from local resources to ignore CacheOnly parameter
Resources that use local storage and have X-Akonadi-Custom-HasLocalStorage=true
entry in their .desktop files are considered 'local'. These resources often have
cacheTimeout set to very low value so that data are not held in Akonadi unnecessary,
because loading them on-demand from the storage is very cheap and fast. However
this causes problem with Baloo, which cannot index content of these items that are
not cached in Akonadi.

For this reason we introduce this hack, that allows FetchHelper to ignore CacheOnly
flag in case all the queries items belong to a local resource. As a result, when Baloo
asks for items from a collection, Akonadi will retrieve payload of these items
from the resources first.

In long term, we need to come up with a better solution, because ItemRetriever
does not really scale when it comes to retrieving many items.

BUG: 332196
FIXED-IN: 1.12.0