Revert all my changes, they only made things worse.
The condition we're waiting on is actually more complex (given that
tasks are not completed when we come back to the top of the loop,
and we want to wait for 1 minute then).

Since this method never releases the lock (except within wait), the race
might not be what I thought it was, anyway.

Sorry for the noise.