Disable global transaction mutex for QSQLITE3 and enable transaction recording
Our QSQLITE3 driver now supports concurrency, so we don't need to serialize
transactions in DataStore anymore. It is however still needed for the
QSQLITE driver shipped with Qt.

Secondary, concurrency support also means possible transactions deadlocks and
timeouts, so we also need to enable transaction recording and replaying for
the QSQLITE3 backend.