ItemSync: Use batch processing and allow throttling
With this patch we process the items in batches and gives resources the
possiblity to throttle the retrieval of messages using the
readyForNextBatch signal. This way we avoid using huge queries and
loading all items into memory, making the whole process a lot less
resource intensive, and make it work for large folder.

REVIEW: 117471