ItemSync: Use serverside merging.
Instead of fetching all local items to check for changes, to then write modified items
back to the server, we use the serverside merging instead.
Additionally this patch allows to merge by gid instead of rid, simply by setting
the gid additionally to the rid.
By using the serverside merging, the code could be simplified and we only need
to fetch the local items anymore to detect removed items when doing a full
listing.

This patch renders ItemSync::updateItem useless, but that should be ok as
I couldn't find a single subclass of ItemSync, and I really don't
expect there to be one out there.