Add a XML file that describes content the database will be populated with
The XML file is transformed into a C++ class called DbPopulator using XSL transformation.

Due to my experience in writing XSL stylesheets being non-existent, the XSL templates are
rather messy and the format and content of the input XML file are limited, but should be
sufficient to populate the database with some testing data that unit-tests can refer to.