Add FakeSearchManager and instantiate an inactive PreprocessorManager from FakeServer
FakeSearchManager is needed to make SearchManager::instance() return a valid
pointer (which is expected by all callers) but not really do anything when
called, as we don't want the search code to interfere with our tests.

Same applies for the PreprocessorManager, but that can be simply turned of and
really does nothing when there are no preprocessor registered.